requirements.txt
openai-whisper==20231117
pyaudio==0.2.14
ffmpeg-python==0.2.0
numpy==1.26.2